# backend/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "v1",
        "title": "My API"
    },
    "paths": {
...
        "/api/QuanLyPhim/LayDanhSachBanner": {
            "get": {
                "tags": [
                    "QuanLyPhim"
                ],
                "operationId": "LayDanhSachBanner",
                "consumes": [],
                "produces": [],
                "parameters": [
                    {
                        "name": "TokenCybersoft",
                        "in": "header",
                        "description": "Nhập token cybersoft",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/api/QuanLyPhim/LayDanhSachPhim": {
            "get": {
                "tags": [
                    "QuanLyPhim"
                ],
                "operationId": "LayDanhSachPhim",
                "consumes": [],
                "produces": [],
                "parameters": [
                    {
                        "name": "maNhom",
                        "in": "query",
                        "required": false,
                        "type": "string",
                        "default": "GP01"
                    },
                    {
                        "name": "tenPhim",
                        "in": "query",
                        "required": false,
                        "type": "string",
                        "default": ""
                    },
                    {
                        "name": "TokenCybersoft",
                        "in": "header",
                        "description": "Nhập token cybersoft",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/api/QuanLyPhim/LayDanhSachPhimPhanTrang": {
            "get": {
                "tags": [
                    "QuanLyPhim"
                ],
                "operationId": "LayDanhSachPhimPhanTrang",
                "consumes": [],
                "produces": [],
                "parameters": [
                    {
                        "name": "maNhom",
                        "in": "query",
                        "required": false,
                        "type": "string",
                        "default": "GP01"
                    },
                    {
                        "name": "tenPhim",
                        "in": "query",
                        "required": false,
                        "type": "string",
                        "default": ""
                    },
                    {
                        "name": "soTrang",
                        "in": "query",
                        "required": false,
                        "type": "integer",
                        "format": "int32",
                        "default": 1
                    },
                    {
                        "name": "soPhanTuTrenTrang",
                        "in": "query",
                        "required": false,
                        "type": "integer",
                        "format": "int32",
                        "default": 10
                    },
                    {
                        "name": "TokenCybersoft",
                        "in": "header",
                        "description": "Nhập token cybersoft",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/api/QuanLyPhim/LayDanhSachPhimTheoNgay": {
            "get": {
                "tags": [
                    "QuanLyPhim"
                ],
                "operationId": "LayDanhSachPhimTheoNgay",
                "consumes": [],
                "produces": [],
                "parameters": [
                    {
                        "name": "maNhom",
                        "in": "query",
                        "required": false,
                        "type": "string",
                        "default": "GP01"
                    },
                    {
                        "name": "tenPhim",
                        "in": "query",
                        "required": false,
                        "type": "string",
                        "default": ""
                    },
                    {
                        "name": "soTrang",
                        "in": "query",
                        "required": false,
                        "type": "integer",
                        "format": "int32",
                        "default": 1
                    },
                    {
                        "name": "soPhanTuTrenTrang",
                        "in": "query",
                        "required": false,
                        "type": "integer",
                        "format": "int32",
                        "default": 10
                    },
                    {
                        "name": "tuNgay",
                        "in": "query",
                        "required": false,
                        "type": "string",
                        "default": ""
                    },
                    {
                        "name": "denNgay",
                        "in": "query",
                        "required": false,
                        "type": "string",
                        "default": ""
                    },
                    {
                        "name": "TokenCybersoft",
                        "in": "header",
                        "description": "Nhập token cybersoft",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/api/QuanLyPhim/ThemPhimUploadHinh": {
            "post": {
                "tags": [
                    "QuanLyPhim"
                ],
                "operationId": "ThemPhimUploadHinh",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [],
                "parameters": [
                    {
                        "name": "frm",
                        "in": "formData",
                        "required": false,
                        "type": "array",
                        "items": {},
                        "collectionFormat": "multi",
                        "uniqueItems": false
                    },
                    {
                        "name": "TokenCybersoft",
                        "in": "header",
                        "description": "Nhập token cybersoft",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/api/QuanLyPhim/CapNhatPhimUpload": {
            "post": {
                "tags": [
                    "QuanLyPhim"
                ],
                "operationId": "CapNhatPhimUpload",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [],
                "parameters": [
                    {
                        "name": "frm",
                        "in": "formData",
                        "required": false,
                        "type": "array",
                        "items": {},
                        "collectionFormat": "multi",
                        "uniqueItems": false
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Nhập token bearer",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "TokenCybersoft",
                        "in": "header",
                        "description": "Nhập token cybersoft",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/api/QuanLyPhim": {
            "post": {
                "tags": [
                    "QuanLyPhim"
                ],
                "operationId": "UploadHinhAnh",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "text/plain",
                    "application/json",
                    "text/json"
                ],
                "parameters": [
                    {
                        "name": "file",
                        "in": "formData",
                        "required": false,
                        "type": "file"
                    },
                    {
                        "name": "tenPhim",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "maNhom",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "TokenCybersoft",
                        "in": "header",
                        "description": "Nhập token cybersoft",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/api/QuanLyPhim/XoaPhim": {
            "delete": {
                "tags": [
                    "QuanLyPhim"
                ],
                "operationId": "XoaPhim",
                "consumes": [],
                "produces": [],
                "parameters": [
                    {
                        "name": "MaPhim",
                        "in": "query",
                        "required": false,
                        "type": "integer",
                        "format": "int32"
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Nhập token bearer",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "TokenCybersoft",
                        "in": "header",
                        "description": "Nhập token cybersoft",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/api/QuanLyPhim/LayThongTinPhim": {
            "get": {
                "tags": [
                    "QuanLyPhim"
                ],
                "operationId": "LayThongTinPhim",
                "consumes": [],
                "produces": [],
                "parameters": [
                    {
                        "name": "MaPhim",
                        "in": "query",
                        "required": false,
                        "type": "integer",
                        "format": "int32",
                        "default": 0
                    },
                    {
                        "name": "TokenCybersoft",
                        "in": "header",
                        "description": "Nhập token cybersoft",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
...
    },
...
}